+++ modules/model_flux.py
@@ -5,7 +5,7 @@ import diffusers
 import transformers
 from safetensors.torch import load_file
 from huggingface_hub import hf_hub_download
-from modules import shared, devices, modelloader, sd_models
+from modules import shared, devices, modelloader, sd_models, sd_unet, model_te
 
 
 debug = shared.log.trace if os.environ.get('SD_LOAD_DEBUG', None) is not None else lambda *args, **kwargs: None
@@ -33,7 +33,7 @@ def load_flux_quanto(checkpoint_info):
         from optimum import quanto # pylint: disable=no-name-in-module
         from optimum.quanto import requantize # pylint: disable=no-name-in-module
     except Exception as e:
-        shared.log.error(f"Loading FLUX: Failed to import optimum-quanto: {e}")
+        shared.log.error(f"Load model: type=FLUX Failed to import optimum-quanto: {e}")
         raise
     quanto.tensor.qbits.QBitsTensor.create = lambda *args, **kwargs: quanto.tensor.qbits.QBitsTensor(*args, **kwargs)
 
@@ -44,7 +44,7 @@ def load_flux_quanto(checkpoint_info):
 
     try:
         quantization_map = os.path.join(repo_path, "transformer", "quantization_map.json")
-        debug(f'Loading FLUX: quantization map="{quantization_map}" repo="{checkpoint_info.name}" component="transformer"')
+        debug(f'Load model: type=FLUX quantization map="{quantization_map}" repo="{checkpoint_info.name}" component="transformer"')
         if not os.path.exists(quantization_map):
             repo_id = sd_models.path_to_repo(checkpoint_info.name)
             quantization_map = hf_hub_download(repo_id, subfolder='transformer', filename='quantization_map.json', cache_dir=shared.opts.diffusers_dir)
@@ -60,16 +60,16 @@ def load_flux_quanto(checkpoint_info):
             try:
                 transformer = transformer.to(dtype=devices.dtype)
             except Exception:
-                shared.log.error(f"Loading FLUX: Failed to cast transformer to {devices.dtype}, set dtype to {transformer.dtype}")
+                shared.log.error(f"Load model: type=FLUX Failed to cast transformer to {devices.dtype}, set dtype to {transformer.dtype}")
     except Exception as e:
-        shared.log.error(f"Loading FLUX: Failed to load Quanto transformer: {e}")
+        shared.log.error(f"Load model: type=FLUX Failed to load Quanto transformer: {e}")
         if debug:
             from modules import errors
             errors.display(e, 'FLUX Quanto:')
 
     try:
         quantization_map = os.path.join(repo_path, "text_encoder_2", "quantization_map.json")
-        debug(f'Loading FLUX: quantization map="{quantization_map}" repo="{checkpoint_info.name}" component="text_encoder_2"')
+        debug(f'Load model: type=FLUX quantization map="{quantization_map}" repo="{checkpoint_info.name}" component="text_encoder_2"')
         if not os.path.exists(quantization_map):
             repo_id = sd_models.path_to_repo(checkpoint_info.name)
             quantization_map = hf_hub_download(repo_id, subfolder='text_encoder_2', filename='quantization_map.json', cache_dir=shared.opts.diffusers_dir)
@@ -87,9 +87,9 @@ def load_flux_quanto(checkpoint_info):
             try:
                 text_encoder_2 = text_encoder_2.to(dtype=devices.dtype)
             except Exception:
-                shared.log.error(f"Loading FLUX: Failed to cast text encoder to {devices.dtype}, set dtype to {text_encoder_2.dtype}")
+                shared.log.error(f"Load model: type=FLUX Failed to cast text encoder to {devices.dtype}, set dtype to {text_encoder_2.dtype}")
     except Exception as e:
-        shared.log.error(f"Loading FLUX: Failed to load Quanto text encoder: {e}")
+        shared.log.error(f"Load model: type=FLUX Failed to load Quanto text encoder: {e}")
         if debug:
             from modules import errors
             errors.display(e, 'FLUX Quanto:')
@@ -122,7 +122,7 @@ def load_flux_bnb(checkpoint_info, diffusers_load_config): # pylint: disable=unu
         else:
             transformer = diffusers.FluxTransformer2DModel.from_single_file(repo_path, **diffusers_load_config)
     except Exception as e:
-        shared.log.error(f"Loading FLUX: Failed to load BnB transformer: {e}")
+        shared.log.error(f"Load model: type=FLUX Failed to load BnB transformer: {e}")
         transformer, text_encoder_2 = None, None
         if debug:
             from modules import errors
@@ -131,7 +131,7 @@ def load_flux_bnb(checkpoint_info, diffusers_load_config): # pylint: disable=unu
 
 
 def load_flux_gguf(file_path): # TODO add support for GGUF flux models
-    shared.log.error(f"Loading FLUX: GGUF UNET is not supported: {file_path}")
+    shared.log.error(f"Load model: type=FLUX GGUF UNET is not supported: {file_path}")
     """
     with torch.device("meta"):
         transformer = diffusers.FluxTransformer2DModel.from_config(os.path.join("configs", "flux", "transformer", "config.json")).to(dtype=devices.dtype)
@@ -182,8 +182,8 @@ def load_transformer(file_path): # triggered by opts.sd_unet change
 def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_checkpoint change
     quant = get_quant(checkpoint_info.path)
     repo_id = sd_models.path_to_repo(checkpoint_info.name)
-    shared.log.debug(f'Loading FLUX: model="{checkpoint_info.name}" repo="{repo_id}" unet="{shared.opts.sd_unet}" t5="{shared.opts.sd_text_encoder}" vae="{shared.opts.sd_vae}" quant={quant} offload={shared.opts.diffusers_offload_mode} dtype={devices.dtype}')
-    debug(f'Loading FLUX: config={diffusers_load_config}')
+    shared.log.debug(f'Load model: type=FLUX model="{checkpoint_info.name}" repo="{repo_id}" unet="{shared.opts.sd_unet}" t5="{shared.opts.sd_text_encoder}" vae="{shared.opts.sd_vae}" quant={quant} offload={shared.opts.diffusers_offload_mode} dtype={devices.dtype}')
+    debug(f'Load model: type=FLUX config={diffusers_load_config}')
     modelloader.hf_login()
 
     transformer = None
@@ -193,8 +193,7 @@ def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_ch
     # load overrides if any
     if shared.opts.sd_unet != 'None':
         try:
-            debug(f'Loading FLUX: unet="{shared.opts.sd_unet}"')
-            from modules import sd_unet
+            debug(f'Load model: type=FLUX unet="{shared.opts.sd_unet}"')
             _transformer = load_transformer(sd_unet.unet_dict[shared.opts.sd_unet])
             if _transformer is not None:
                 transformer = _transformer
@@ -202,14 +201,14 @@ def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_ch
                 shared.opts.sd_unet = 'None'
                 sd_unet.failed_unet.append(shared.opts.sd_unet)
         except Exception as e:
-            shared.log.error(f"Loading FLUX: Failed to load UNet: {e}")
+            shared.log.error(f"Load model: type=FLUX Failed to load UNet: {e}")
             shared.opts.sd_unet = 'None'
             if debug:
                 from modules import errors
                 errors.display(e, 'FLUX UNet:')
     if shared.opts.sd_text_encoder != 'None':
         try:
-            debug(f'Loading FLUX: t5="{shared.opts.sd_text_encoder}"')
+            debug(f'Load model: type=FLUX t5="{shared.opts.sd_text_encoder}"')
             from modules.model_te import load_t5
             _text_encoder_2 = load_t5(name=shared.opts.sd_text_encoder, cache_dir=shared.opts.diffusers_dir)
             if _text_encoder_2 is not None:
@@ -217,14 +216,14 @@ def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_ch
             else:
                 shared.opts.sd_text_encoder = 'None'
         except Exception as e:
-            shared.log.error(f"Loading FLUX: Failed to load T5: {e}")
+            shared.log.error(f"Load model: type=FLUX Failed to load T5: {e}")
             shared.opts.sd_text_encoder = 'None'
             if debug:
                 from modules import errors
                 errors.display(e, 'FLUX T5:')
     if shared.opts.sd_vae != 'None' and shared.opts.sd_vae != 'Automatic':
         try:
-            debug(f'Loading FLUX: vae="{shared.opts.sd_vae}"')
+            debug(f'Load model: type=FLUX vae="{shared.opts.sd_vae}"')
             from modules import sd_vae
             # vae = sd_vae.load_vae_diffusers(None, sd_vae.vae_dict[shared.opts.sd_vae], 'override')
             vae_file = sd_vae.vae_dict[shared.opts.sd_vae]
@@ -232,7 +231,7 @@ def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_ch
                 vae_config = os.path.join('configs', 'flux', 'vae', 'config.json')
                 vae = diffusers.AutoencoderKL.from_single_file(vae_file, config=vae_config, **diffusers_load_config)
         except Exception as e:
-            shared.log.error(f"Loading FLUX: Failed to load VAE: {e}")
+            shared.log.error(f"Load model: type=FLUX Failed to load VAE: {e}")
             shared.opts.sd_vae = 'None'
             if debug:
                 from modules import errors
@@ -248,7 +247,7 @@ def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_ch
             if _text_encoder is not None:
                 text_encoder_2 = _text_encoder
         except Exception as e:
-            shared.log.error(f"Loading FLUX: Failed to load NF4 components: {e}")
+            shared.log.error(f"Load model: type=FLUX Failed to load NF4 components: {e}")
             if debug:
                 from modules import errors
                 errors.display(e, 'FLUX NF4:')
@@ -260,7 +259,7 @@ def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_ch
             if _text_encoder is not None:
                 text_encoder_2 = _text_encoder
         except Exception as e:
-            shared.log.error(f"Loading FLUX: Failed to load Quanto components: {e}")
+            shared.log.error(f"Load model: type=FLUX Failed to load Quanto components: {e}")
             if debug:
                 from modules import errors
                 errors.display(e, 'FLUX Quanto:')
@@ -269,11 +268,13 @@ def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_ch
     components = {}
     if transformer is not None:
         components['transformer'] = transformer
+        sd_unet.loaded_unet = shared.opts.sd_unet
     if text_encoder_2 is not None:
         components['text_encoder_2'] = text_encoder_2
+        model_te.loaded_te = shared.opts.sd_text_encoder
     if vae is not None:
         components['vae'] = vae
-    shared.log.debug(f'Loading FLUX: preloaded={list(components)}')
+    shared.log.debug(f'Load model: type=FLUX preloaded={list(components)}')
     if repo_id == 'sayakpaul/flux.1-dev-nf4':
         repo_id = 'black-forest-labs/FLUX.1-dev' # workaround since sayakpaul model is missing model_index.json
     pipe = diffusers.FluxPipeline.from_pretrained(repo_id, cache_dir=shared.opts.diffusers_dir, **components, **diffusers_load_config)